refactor(solver): split display alias backref helpers
